Fix bugs found in production readiness review
Must-fix:
- Fix from_pretrained kwarg: dtype= -> torch_dtype= in _full.py
  (wrong kwarg name could cause model to load in wrong precision)
- Add expert count mismatch validation in expand_expert_lora_tensors
  (mismatched counts would silently produce wrong results)

Should-fix:
- Reorder ShardWriter.flush() to increment shard count only after
  successful save_file (avoids inconsistent state on disk-full errors)
- Use base_model.config.to_dict() instead of fragile class name string
  matching in merge_adapter_weights (works with wrapped models)

Tests added:
- Expert count mismatch raises ValueError
- unembed_tokens -> lm_head remapping

+++ tinker_cookbook/weights/_export/_full.py
@@ -76,5 +76,5 @@ def _load_model(
 ) -> PreTrainedModel:
     auto_cls = AutoModelForImageTextToText if is_multimodal(config) else AutoModelForCausalLM
     return auto_cls.from_pretrained(
-        model_path, dtype=torch_dtype, trust_remote_code=trust_remote_code
+        model_path, torch_dtype=torch_dtype, trust_remote_code=trust_remote_code
     )
